test: split TTS characterization fixtures
